auth-service/requirements.txt
fastapi==0.115.12
uvicorn==0.21.0
sqlalchemy==2.0.0
pyjwt==2.6.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
fastapi-mail==1.4.2
alembic==1.9.0
psycopg2-binary==2.9.5
pydantic_settings==2.8.1
pydantic==2.10.6
python-dotenv==1.0.0
email-validator==2.2.0
python-jose[cryptography]==3.3.0
